Resorted to using a single ViewModel
