+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/SingleLiveEvent.kt
@@ -16,11 +16,11 @@ package de.rki.coronawarnapp.ui
  *  limitations under the License.
  */
 
-import android.util.Log
 import androidx.annotation.MainThread
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
+import timber.log.Timber
 import java.util.concurrent.atomic.AtomicBoolean
 
 /**
@@ -40,7 +40,8 @@ class SingleLiveEvent<T> : MutableLiveData<T>() {
     @MainThread
     override fun observe(owner: LifecycleOwner, observer: Observer<in T>) {
         if (hasActiveObservers()) {
-            Log.w(TAG, "Multiple observers registered but only one will be notified of changes.")
+            Timber
+                .w("Multiple observers registered but only one will be notified of changes.")
         }
         // Observe the internal MutableLiveData
         super.observe(owner, Observer { t ->
@@ -61,8 +62,5 @@ class SingleLiveEvent<T> : MutableLiveData<T>() {
     @MainThread
     fun call() {
         value = null
-    }
-    companion object {
-        private val TAG = "SingleLiveEvent"
     }
 }
