+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/submission/EuropeanConsentEvent.kt
@@ -2,6 +2,6 @@ package de.rki.coronawarnapp.ui.submission
 
 sealed class EuropeanConsentEvent {
     object NavigateToTargetGermany : EuropeanConsentEvent()
-    object NavigateToPreviousScreen: EuropeanConsentEvent()
-    object NavigateToKeysSubmission: EuropeanConsentEvent()
+    object NavigateToPreviousScreen : EuropeanConsentEvent()
+    object NavigateToKeysSubmission : EuropeanConsentEvent()
 }
