+++ Corona-Warn-App/src/main/res/drawable-night/ic_submission_illustration_european_consent.xml
