+++ Corona-Warn-App/src/main/res/drawable/ic_submission_illustration_european_consent.xml
