+++ Corona-Warn-App/src/main/res/values/strings.xml
@@ -619,12 +619,12 @@
     <string name="settings_background_priority_card_button">"Open Device Settings"</string>
     <!-- XHED : settings(background priority) - headline on card about the current status and what to do -->
     <string name="settings_background_priority_card_headline">"Disable prioritized background activity"</string>
-    <!-- XTXT: settings(EFGS consent) - status next to switch under title -->
-    <string name="settings_efgs_consent_status_active">"Activated Dummy Text"</string>
-    <!-- XTXT: settings(EFGS consent) - status next to switch under title -->
-    <string name="settings_efgs_consent_status_inactive">"Stopped Dummy Text"</string>
-    <!-- XHED: settings(Efgs consent) - page title -->
-    <string name="settings_efgs_consent_title">"European Data Consent Dummy Text"</string>
+    <!-- XTXT: settings(european federal gateway server consent) - status next to switch under title -->
+    <string name="settings_european_consent_status_active" />
+    <!-- XTXT: settings(european federal gateway server consent) - status next to switch under title -->
+    <string name="settings_european_consent_status_inactive" />
+    <!-- XHED: settings(european federal gateway server consent) - page title -->
+    <string name="settings_european_consent_title" />
 
     <!-- ####################################
                App Information
